makes some slight changes in style and removes compiler warnings. code could use a lot of DRY, and the project should be set up in multiple folders because all of these are main functions. you learn a lot in 3 years

+++ primes.c
@@ -25,7 +25,7 @@ bool IsItPrime(int candidate)
 		return false;
 	else
 	{
-		last = sqrt(candidate);
+		last = (int)sqrt(candidate);
 
 		for (i = 2; i <= last; i++)
 		{
